Add images,recipes folders
For each recipe, an image is selected from the images folder to be used in the project. All recipes are linked from the main index.html file which links to the appropriate recipe. At the main index page, each recipe is linked using links. Whenever recipe link is clicked on the index.html, it opens the appropriate recipe.Each recipe has a description about the recipe,Ingredients required to cook the recipe and detailed step by step guide to help during cooking that specified recipe.